zoraasi/requirements.txt
ijson>=3.2.0
openai>=1.0.0
requests>=2.28.0
pymupdf>=1.23.0
python-docx>=1.0.0
